Render cells within the usual component lifecycle
Previously, the first render pass would create
empty shells for cell content, but the actual
cell content (or placeholders) would subsequently
be rendered into those outside of the usual
Lit lifecycle.

However, this causes a number of undesired
behaviors, including a resistance to SSR
(which requires all content to be rendered
on the first pass regardless of visibility)
and unusual behavior when the content needs
to be reloaded (flickering during the empty
shell limbo, and manual effort required
by the consumer to ensure prior pages of
content becoming visible actually triggers
the right events).

With this update, we keep the scroll-optimized
behavior while making the render cycles
much more ordinary for a Lit component,
which should make the infinite scroller
easier to work with overall.